backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pydantic==2.8.2
pydantic-settings==2.0.3
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
websockets==12.0
python-socketio==5.10.0
